+++ lib/mui/color_scheme.rb
@@ -28,6 +28,10 @@ module Mui
       syntax_global_variable
       syntax_method_call
       syntax_type
+      diff_add
+      diff_delete
+      diff_hunk
+      diff_header
     ].freeze
 
     attr_reader :name, :colors